FYP_Model/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
tensorflow-cpu==2.16.1
